refactor: address review feedback on token auth (#20)
- Declare uvicorn as an explicit dependency instead of relying on it
  transitively through fastmcp
- Move BearerTokenMiddleware into its own module (middleware.py) so
  server.py stays focused on CLI parsing and routing
- Poll uvicorn.Server.started in the live test instead of sleeping a
  fixed interval (faster locally, no flaky CI timeouts)
